refactor: split system.go into subscription and publisher files
