+++ master/internal/stream/messages.go
@@ -1,0 +1,43 @@
+package stream
+
+import (
+	"encoding/json"
+
+	"github.com/gorilla/websocket"
+	log "github.com/sirupsen/logrus"
+
+	"github.com/determined-ai/determined/master/pkg/stream"
+)
+
+// StartupMsg is the first message a streaming client sends.
+//
+// It declares initially known keys and also configures the initial subscriptions for the stream.
+type StartupMsg struct {
+	SyncID    string              `json:"sync_id"`
+	Known     KnownKeySet         `json:"known"`
+	Subscribe SubscriptionSpecSet `json:"subscribe"`
+}
+
+// KnownKeySet allows a client to describe which primary keys it knows of as existing, so the server
+// can respond with a different KnownKeySet of deleted messages of client-known keys that don't
+// exist.
+//
+// Each field of a KnownKeySet is a comma-separated list of int64s and ranges like "a,b-c,d".
+type KnownKeySet struct {
+	Projects string `json:"projects"`
+}
+
+// prepareWebsocketMessage converts the PreparableMessage into a websocket.PreparedMessage.
+func prepareWebsocketMessage(obj stream.PreparableMessage) interface{} {
+	jbytes, err := json.Marshal(obj)
+	if err != nil {
+		log.Errorf("error marshaling message for streaming: %s", err.Error())
+		return nil
+	}
+	msg, err := websocket.NewPreparedMessage(websocket.TextMessage, jbytes)
+	if err != nil {
+		log.Errorf("error preparing message for streaming: %s", err.Error())
+		return nil
+	}
+	return msg
+}
